requirements.txt
pyTelegramBotAPI
Flask
requests
gTTS
pillow
beautifulsoup4
pydub